Add files via upload
Jupyter Notebook and Power BI files featuring an analysis of the quality of surface observations, including the identification of missing data and outliers, the implementation of unit tests for data quality assurance, and solutions for imputing missing values. The data comprises hourly weather observations from the Met Office for 11 August 2024.